+++ crates/tno-core/src/error.rs
@@ -6,10 +6,13 @@ use crate::runner::RunnerError;
 pub enum CoreError {
     #[error("no suitable runner for task kind: {0}")]
     NoRunner(String),
+
     #[error("supervisor error: {0}")]
     Supervisor(String),
+
     #[error("mapping error: {0}")]
     Mapping(String),
+
     #[error("runner error: {0}")]
     Runner(#[from] RunnerError),
 }
