--- crates/tno-core/src/map/backoff.rs
@@ -1,16 +1,0 @@
-use std::time::Duration;
-
-use taskvisor::BackoffPolicy;
-use tno_model::BackoffStrategy;
-
-use super::to_jitter_policy;
-
-pub fn to_backoff_policy(s: &BackoffStrategy) -> BackoffPolicy {
-    BackoffPolicy {
-        success_delay: s.delay_ms.map(Duration::from_millis),
-        first: Duration::from_millis(s.first_ms),
-        max: Duration::from_millis(s.max_ms),
-        jitter: to_jitter_policy(s.jitter),
-        factor: s.factor,
-    }
-}
