+++ crates/tno-core/src/map/mod.rs
@@ -1,11 +1,6 @@
-mod admission;
-mod backoff;
-mod jitter;
-mod restart;
-mod spec;
+//! Adapter layer between `tno-model` (public specs) and the taskvisor runtime.
+//!
+//! This crate maps high-level API types into taskvisor’s internal execution structures.
 
-pub use admission::to_admission_policy;
-pub use backoff::to_backoff_policy;
-pub use jitter::to_jitter_policy;
-pub use restart::to_restart_policy;
-pub use spec::{to_controller_spec, to_task_spec};
+mod mapping;
+pub use mapping::*;
